Transform task to avoid axioms for goals in Grounder
If the goal is not a conjunction of literals, Fast Downward introduces
axioms to handle it. To avoid this, we instead introduce an artificial
goal action (removed when mapping back).

There can be several ground representatives of the artificial goal
action in the compiled task (e.g. from disjunctive preconditions).
We now give each instance an individual name.